stuff and other stuff

+++ build/android/build.py
@@ -47,11 +47,11 @@ if len(errlines) == 0:
       if f.endswith('.moon'):
         handleMoonFile(os.path.join(root, f))
 
-  os.system("ant debug")
-  adb = "~/stuff/work/android/adt-bundle-linux-x86_64-20140702/sdk/platform-tools/adb "
-  os.system(adb+" -d install -r bin/FrameworkTest-debug.apk")
-  os.system(adb+" -d shell am kill com.spacekomodo.berrybounce/.MainActivity")
-  os.system(adb+" -d shell am start -n com.spacekomodo.berrybounce/.MainActivity")
+  if os.system("ant debug") == 0:
+    adb = "~/stuff/work/android/adt-bundle-linux-x86_64-20140702/sdk/platform-tools/adb "
+    os.system(adb+" -d install -r bin/FrameworkTest-debug.apk")
+    os.system(adb+" -d shell am kill com.spacekomodo.berrybounce/.MainActivity")
+    os.system(adb+" -d shell am start -n com.spacekomodo.berrybounce/.MainActivity")
 else:
   for e in errlines:
     fileName = e[:e.find(":")]
